#71 Cut down code to be just in inventory.py, deleted itemMenu.py, moved classes and helpers to invClassHelpers.py for readability
